feat: 부서 및 위치 정보 조회, 특정 부서 급여를 특정 비율로 인상

+++ src/main/kotlin/com/project/ecoNrich/domain/employee/entity/EmployeeDetailView.kt
@@ -23,7 +23,7 @@ class EmployeeDetailView(
     val departmentId: Int? = null,
 
     @Column(name = "location_id")
-    val locationId: Int? = null,
+    val locationId: Long? = null,
 
     @Column(name = "country_id")
     val countryId: String,
